+++ packages/atomic/scripts/externalPackageMappings.ts
@@ -2,8 +2,15 @@ import path from 'node:path';
 import buenoJson from '../../bueno/package.json';
 import headlessJson from '../../headless/package.json';
 
-const headlessVersion = 'v' + headlessJson.version;
-const buenoVersion = 'v' + buenoJson.version;
+const isNightly = process.env.IS_NIGHTLY === 'true';
+
+const headlessVersion = isNightly
+  ? `v${headlessJson.version.split('.').shift()}-nightly`
+  : 'v' + headlessJson.version;
+
+const buenoVersion = isNightly
+  ? `v${buenoJson.version.split('.').shift()}-nightly`
+  : 'v' + buenoJson.version;
 
 export function generateExternalPackageMappings(basePath: string): {
   [key: string]: {devWatch: string; cdn: string};