+++ packages/quantic/force-app/main/default/lwc/quanticDateFacet/__tests__/quanticDateFacet.test.js
@@ -47,7 +47,10 @@ function createTestComponent(options = defaultOptions) {
 
 const functionsMocks = {
   buildDateFacet: jest.fn(() => dateFacetControllerMock),
-  buildFacetConditionsManager: jest.fn(),
+  stopWatching: jest.fn(),
+  buildFacetConditionsManager: jest.fn(() => ({
+    stopWatching: functionsMocks.stopWatching,
+  })),
   buildSearchStatus: jest.fn(() => ({
     subscribe: jest.fn((callback) => callback()),
     state: {},
@@ -187,6 +190,26 @@ describe('c-quantic-date-facet', () => {
 
     afterAll(() => {
       functionsMocks.buildDateFacet.mockReturnValue(dateFacetControllerMock);
+    });
+  });
+
+  describe('when the component is disconnected', () => {
+    it('should make the condition manager stop watching the facet', async () => {
+      const exampleFacetDependency = {
+        parentFacetId: 'filetype',
+        expectedValue: 'txt',
+      };
+      const element = createTestComponent({
+        ...defaultOptions,
+        dependsOn: exampleFacetDependency,
+      });
+      await flushPromises();
+      expect(functionsMocks.buildFacetConditionsManager).toHaveBeenCalledTimes(
+        1
+      );
+
+      document.body.removeChild(element);
+      expect(functionsMocks.stopWatching).toHaveBeenCalledTimes(1);
     });
   });
 });